Transfers package update logic to update_package_task method
